pyproject.toml
[tool.poetry]
name = "codeforher-backend"
version = "0.1.0"
description = ""
authors = ["Ashish K <64360092+Ashishkumaraswamy@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.11"
pymongo = "^4.11.3"
bcrypt = "^4.3.0"
python-dotenv = "^1.0.1"
pyjwt = "^2.10.1"
black = "^25.1.0"
isort = "^6.0.1"
uvicorn = "^0.34.0"
structlog = "^25.2.0"
email-validator = "^2.2.0"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
twilio = "^9.5.1"
geopy = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
